refactor unify common types
